Moved common template functions into a utils module
